Added Loss Function
Categorical Entropy Loss Function
Average Loss per Batch

+++ src/utils.hpp
@@ -4,7 +4,7 @@
 namespace utils {
 	double random(const double&, const double&);
 
-	std::tuple<nc::NdArray<double>, nc::NdArray<double>> spiral_data(const size_t&, const size_t&);
+	std::tuple<nc::NdArray<double>, nc::NdArray<int>> spiral_data(const size_t&, const size_t&);
 
 	nc::NdArray<double> addVectorToEveryRow(const nc::NdArray<double>&, const nc::NdArray<double>&);
